+++ entities/src/bpmn_20_call_conversation.rs
@@ -7,11 +7,11 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_call_conversation")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : ConversationNode
-    pub super_conversation_node: i32,
+    pub super_conversation_node: i64,
     /// COMPLEX FIELD : CallConversation-calledCollaborationRef
-    pub called_collaboration_ref: Option<i32>,
+    pub called_collaboration_ref: Option<i64>,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
