+++ entities/src/bpmn_20_catch_event.rs
@@ -7,11 +7,11 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_catch_event")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : Event
-    pub super_event: i32,
+    pub super_event: i64,
     /// COMPLEX FIELD : CatchEvent-outputSet
-    pub output_set: Option<i32>,
+    pub output_set: Option<i64>,
     /// SIMPLE FIELD : CatchEvent-parallelMultiple
     pub parallel_multiple: std::primitive::bool,
 }
