+++ entities/src/bpmn_20_choreography.rs
@@ -7,11 +7,11 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_choreography")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : FlowElementsContainer
-    pub super_flow_elements_container: i32,
+    pub super_flow_elements_container: i64,
     /// SUPER FIELD : Collaboration
-    pub super_collaboration: i32,
+    pub super_collaboration: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
