+++ entities/src/bpmn_20_conversation_association.rs
@@ -7,13 +7,13 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_conversation_association")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : BaseElement
-    pub super_base_element: i32,
+    pub super_base_element: i64,
     /// COMPLEX FIELD : ConversationAssociation-innerConversationNodeRef
-    pub inner_conversation_node_ref: i32,
+    pub inner_conversation_node_ref: i64,
     /// COMPLEX FIELD : ConversationAssociation-outerConversationNodeRef
-    pub outer_conversation_node_ref: i32,
+    pub outer_conversation_node_ref: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
