+++ entities/src/bpmn_20_conversation_node.rs
@@ -7,11 +7,11 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_conversation_node")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : InteractionNode
-    pub super_interaction_node: i32,
+    pub super_interaction_node: i64,
     /// SUPER FIELD : BaseElement
-    pub super_base_element: i32,
+    pub super_base_element: i64,
     /// SIMPLE FIELD : ConversationNode-name
     pub name: std::string::String,
 }
