+++ entities/src/bpmn_20_documentation.rs
@@ -7,9 +7,9 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_documentation")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : BaseElement
-    pub super_base_element: i32,
+    pub super_base_element: i64,
     /// SIMPLE FIELD : Documentation-text
     pub text: std::string::String,
     /// SIMPLE FIELD : Documentation-textFormat
