+++ entities/src/bpmn_20_group.rs
@@ -7,11 +7,11 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "bpmn_20_group")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     /// SUPER FIELD : Artifact
-    pub super_artifact: i32,
+    pub super_artifact: i64,
     /// COMPLEX FIELD : Group-categoryValueRef
-    pub category_value_ref: Option<i32>,
+    pub category_value_ref: Option<i64>,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
